+++ src/constants/appVersion.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = "13.213.30";
+export const APP_VERSION = "13.213.31";