+++ src/hooks/shared/index.ts
@@ -11,3 +11,4 @@ export * from './useTabsParam';
 export * from './useToast';
 export * from './useIsMobile';
 export * from './useRadixPointerEventsRescue';
+export * from './useColumnVisibility';